bug 1462749: move site status page to django admin
