+++ webapp-django/crashstats/manage/admin_urls.py
@@ -7,4 +7,8 @@ urlpatterns = [
     url('^crash-me-now/$',
         admin.crash_me_now,
         name='crash_me_now'),
+    url('^sitestatus/$',
+        admin.site_status,
+        name='site_status'),
+
 ]
