+++ webapp-django/crashstats/manage/jinja2/manage/home.html
@@ -27,10 +27,6 @@
 
         <h3>Tools</h3>
         <ul>
-            <li><a href="{{ url('manage:site_status') }}" >
-                <span class="fa fa-check-circle fa-5x"></span>
-                Site Status
-            </a></li>
             <li><a href="{{ url('manage:analyze_model_fetches') }}">
                 <span class="fa fa-dashboard fa-5x"></span>
                 Analyze Model Fetches
