--- webapp-django/crashstats/manage/jinja2/manage/site_status.html
@@ -1,73 +1,0 @@
-{% extends "manage/base.html" %}
-
-{% block page_title %}{{ super() }} - Site Status{% endblock %}
-
-{% block admin_title %}{{ super() }} - Site Status{% endblock %}
-
-{% block mainbody %}
-<div class="panel">
-    <div class="body notitle">
-        <h3>Version info</h3>
-        {% for url, info in version_info.items() %}
-            <p>{{ url }}: {{ info }}</p>
-        {% endfor %}
-
-        <h3>Alembic status</h3>
-        <p>
-            {% if alembic_error %}
-                {{ alembic_error }}
-            {% else %}
-                HEAD: {{ alembic_version }}
-            {% endif %}
-        </p>
-
-        <h3>Django db status</h3>
-        {% if django_error %}
-            <p>{{ django_error }}</p>
-        {% else %}
-            <table class="data-table">
-                <tr>
-                    <th>id</th>
-                    <th>app</th>
-                    <th>name</th>
-                    <th>applied</th>
-                </tr>
-                {% for row in django_db_data %}
-                    <tr>
-                        <td>{{ row['id'] }}</td>
-                        <td>{{ row['app'] }}</td>
-                        <td>{{ row['name'] }}</td>
-                        <td>{{ row['applied'] }}</td>
-                    </tr>
-                {% endfor %}
-            </table>
-        {% endif %}
-
-        <h3>Crontabber failed jobs</h3>
-        {% if crontabber_error %}
-            <p>{{ crontabber_error }}</p>
-        {% else %}
-            <table class="data-table">
-                <tr>
-                    <th>app_name</th>
-                    <th>next_run</th>
-                    <th>last_run</th>
-                    <th>last_success</th>
-                    <th>error_count</th>
-                    <th>last_error</th>
-                </tr>
-                {% for row in crontabber_data %}
-                    <tr>
-                        <td>{{ row['app_name'] }}</td>
-                        <td>{{ row['next_run'] }}</td>
-                        <td>{{ row['last_run'] }}</td>
-                        <td>{{ row['last_success'] }}</td>
-                        <td>{{ row['error_count'] }}</td>
-                        <td>{{ row['last_error'] }}</td>
-                    </tr>
-                {% endfor %}
-            </table>
-        {% endif %}
-    </div>
-</div>
-{% endblock %}
