+++ webapp-django/crashstats/manage/templates/admin/index.html
@@ -27,6 +27,13 @@
                 <td></td>
                 <td></td>
             </tr>
+            <tr>
+                <th scope="row">
+                    <a href="{% url 'siteadmin:site_status' %}">Site status</a>
+                </th>
+                <td></td>
+                <td></td>
+            </tr>
         </table>
     </div>
     {# END SECTION #}
