+++ webapp-django/crashstats/manage/tests/test_admin_views.py
@@ -5,9 +5,9 @@ from crashstats.crashstats.tests.test_views import BaseTestViews
 import pytest
 
 
-class TestCrashMeNow(BaseTestViews):
+class SiteAdminTestViews(BaseTestViews):
     def _login(self, is_superuser=True):
-        user = super(TestCrashMeNow, self)._login(
+        user = super(SiteAdminTestViews, self)._login(
             username='lonnen',
             email='lonnen@example.com',
         )
@@ -16,6 +16,8 @@ class TestCrashMeNow(BaseTestViews):
         user.save()
         return user
 
+
+class TestCrashMeNow(SiteAdminTestViews):
     def test_view(self):
         url = reverse('siteadmin:crash_me_now')
         response = self.client.get(url)
@@ -24,3 +26,15 @@ class TestCrashMeNow(BaseTestViews):
         self._login()
         with pytest.raises(ZeroDivisionError):
             self.client.get(url)
+
+
+class TestSiteStatus(SiteAdminTestViews):
+    def test_page_load(self):
+        """Basic test to make sure the page loads at all"""
+        url = reverse('siteadmin:site_status')
+        response = self.client.get(url)
+        assert response.status_code == 302
+
+        self._login()
+        response = self.client.get(url)
+        assert response.status_code == 200
